+++ pkgs/gfx/catimg.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -44,7 +45,7 @@ int main(int argc, char **argv) {
     }
     struct fbdev_bitblit bitblit = {
         .target_buffer = GFX_FRONT_BUFFER,
-        .source = (unsigned long*)data,
+        .source = (uint32_t *)data,
         .x = 0,
         .y = 0,
         .width = w,
